Spells: Fixed Command Shout removing some health when refreshing
Former-commit-id: 156daa6370bc6cb7d0a9befc75c2fc0cfaa2a017

+++ src/server/game/AI/SmartScript.cpp
@@ -1069,7 +1069,7 @@ void SmartScript::ProcessAction(SmartScriptHolder& e, Unit* unit, uint32 var0, u
                         continue;
 
                     if (e.action.invincHP.percent)
-                        ai->SetInvincibilityHpLevel(target->ToCreature()->CountPctFromMaxHealth(e.action.invincHP.percent));
+                        ai->SetInvincibilityHpLevel(target->ToCreature()->CountPctFromMaxHealth(int32(e.action.invincHP.percent)));
                     else
                         ai->SetInvincibilityHpLevel(e.action.invincHP.minHP);
                 }
